Migrating to Sf 6.3 and PHP 8.1+, major cleanup and major compatibility breaks

+++ Doctrine/DoctrineHelper.php
@@ -2,7 +2,7 @@
 /*
  * This file is part of the Sidus/AdminBundle package.
  *
- * Copyright (c) 2015-2021 Vincent Chalnot
+ * Copyright (c) 2015-2023 Vincent Chalnot
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -16,30 +16,28 @@ use Doctrine\Common\Util\ClassUtils;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\ManagerRegistry;
 use InvalidArgumentException;
-use Sidus\AdminBundle\Admin\Action;
-use Sidus\BaseBundle\Translator\TranslatableTrait;
-use Symfony\Component\HttpFoundation\Session\Session;
+use Sidus\AdminBundle\Session\FlashHelper;
+use Sidus\AdminBundle\Model\Action;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
-use Symfony\Contracts\Translation\TranslatorInterface;
 
 /**
  * Provides a simple way to access Doctrine utilities from a controller or an action
  */
 class DoctrineHelper
 {
-    use TranslatableTrait;
-
     public function __construct(
-        protected ManagerRegistry $doctrine,
-        TranslatorInterface $translator,
+        protected FlashHelper $flashHelper,
+        protected ?ManagerRegistry $managerRegistry = null,
     ) {
-        $this->translator = $translator;
     }
 
     public function getManagerForEntity(object $entity): EntityManagerInterface
     {
+        if (!$this->managerRegistry) {
+            throw new InvalidArgumentException('Doctrine is not enabled for Sidus/AdminBundle');
+        }
         $class = ClassUtils::getClass($entity);
-        $entityManager = $this->doctrine->getManagerForClass($class);
+        $entityManager = $this->managerRegistry->getManagerForClass($class);
         if (!$entityManager instanceof EntityManagerInterface) {
             throw new InvalidArgumentException("No manager found for class {$class}");
         }
@@ -49,36 +47,44 @@ class DoctrineHelper
 
     public function saveEntity(Action $action, object $entity, SessionInterface $session = null): void
     {
+        if (!$this->managerRegistry) {
+            throw new InvalidArgumentException('Doctrine is not enabled for Sidus/AdminBundle');
+        }
         $entityManager = $this->getManagerForEntity($entity);
         $entityManager->persist($entity);
         $entityManager->flush();
 
-        $this->addFlash($action, $session);
+        $this->flashHelper->addFlash($action, $session);
     }
 
     public function deleteEntity(Action $action, object $entity, SessionInterface $session = null): void
     {
+        if (!$this->managerRegistry) {
+            throw new InvalidArgumentException('Doctrine is not enabled for Sidus/AdminBundle');
+        }
         $entityManager = $this->getManagerForEntity($entity);
         $entityManager->remove($entity);
         $entityManager->flush();
 
-        $this->addFlash($action, $session);
+        $this->flashHelper->addFlash($action, $session);
     }
 
-    public function addFlash(Action $action, SessionInterface $session = null): void
+    public function entityToString(object $data): string
     {
-        if ($action && $session instanceof Session) {
-            $session->getFlashBag()->add(
-                'success',
-                $this->tryTranslate(
-                    [
-                        "sidus.admin.{$action->getAdmin()->getCode()}.{$action->getCode()}.success",
-                        "sidus.admin.flash.{$action->getCode()}.success",
-                    ],
-                    [],
-                    ucfirst($action->getCode()).' success'
-                )
-            );
+        if (!$this->managerRegistry) {
+            throw new InvalidArgumentException('Doctrine is not enabled for Sidus/AdminBundle');
         }
+        $manager = $this->managerRegistry->getManagerForClass($data::class);
+        if (!$manager instanceof EntityManagerInterface) {
+            throw new InvalidArgumentException('Not a Doctrine entity');
+        }
+
+        $classMetadata = $manager->getClassMetadata($data::class);
+        $identifiers = $classMetadata->getIdentifierValues($data);
+        if (1 === count($identifiers)) {
+            return $data::class.'#'.reset($identifiers);
+        }
+
+        return $data::class.json_encode($identifiers, JSON_THROW_ON_ERROR);
     }
 }
